Simple: introduce the concept of toolbox
As an Haruspex may need more than a debugger to inspect a crash archive,
introduce the concept of toolbox, i.e. the collection of tools needed by
Haruspex, declared in the preset, implemented as a dictionary.

So, add the adequate method to create this toolbox from preset and
update the constructor of SimpleCoreHaruspex accordingly.